+++ src/main/php/com/google/authenticator/SecretBytes.class.php
@@ -7,8 +7,7 @@ use lang\types\Bytes;
  *
  * @test xp://com.google.authenticator.unittest.SecretBytesTest
  */
-class SecretBytes extends \lang\Object implements Secret {
-  private $bytes;
+class SecretBytes extends Secret {
 
   /**
    * Creates a new secret string given base32-encoded bytes
@@ -17,24 +16,6 @@ class SecretBytes extends \lang\Object implements Secret {
    * @throws lang.FormatException
    */
   public function __construct($bytes) {
-    $this->bytes= $bytes instanceof Bytes ? $bytes->buffer : $bytes;
-  }
-
-  /** @return string */
-  public function bytes() { return $this->bytes; }
-
-  /**
-   * Creates a string representation of this secret. Yields as many
-   * asterisks as there are bytes in this secret.
-   *
-   * @return string
-   */
-  public function toString() {
-    return $this->getClassName().'('.str_repeat('*', strlen($this->bytes)).')';
-  }
-
-  /** @return [:var] */
-  public function __debugInfo() {
-    return ['bytes' => str_repeat('*', strlen($this->bytes))];
+    parent::__construct($bytes instanceof Bytes ? $bytes->buffer : $bytes);
   }
 }