+++ ChangeLog.md
@@ -3,6 +3,9 @@ Google Authenticator ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Implemented `Secret::random()` as requested and drafted in issue #2.
+  (@thekid)
+
 ## 0.1.0 / 2014-12-21
 
 * Hello World! First release - @thekid